fix: auto enable agent sync workload-v

+++ server/controller/trisolaris/config/config.go
@@ -39,8 +39,8 @@ type Config struct {
 	Chrony                         Chrony   `yaml:"chrony"`
 	SelfUpdateUrl                  string   `default:"grpc" yaml:"self-update-url"`
 	RemoteApiTimeout               uint16   `default:"30" yaml:"remote-api-timeout"`
-	TridentTypeForUnknowVtap       uint16   `default:"0" yaml:"trident-type-for-unknow-vtap"`
-	OldTridentTypeForUnknowVtap    uint16   `default:"0" yaml:"trident-type-for-unkonw-vtap"` // Compatible with older configurations
+	TridentTypeForUnknowVtap       int      `default:"0" yaml:"trident-type-for-unknow-vtap"`
+	OldTridentTypeForUnknowVtap    int      `default:"0" yaml:"trident-type-for-unkonw-vtap"` // Compatible with older configurations
 	PlatformVips                   []string `yaml:"platform-vips"`
 	NodeType                       string   `default:"master" yaml:"node-type"`
 	RegionDomainPrefix             string   `yaml:"region-domain-prefix"`
